Update LICENSE.md
Should not change too much for us, the main thing is that people who use our code are not required to use the same license. The goal is still the same as before, anyone can use and sell including without attribution, this removes friction from people wanting to do so and prevents cases where it would technically be illigal to use our stuff if not all derived code can be same licensed which I don't want. I'll admit that was kind of my main goal but I chose gpl3 without too much research originally.

tldr: our code can be used in more places with less friction